requirements.txt
PyAutoGUI~=0.9.53
Pillow~=9.0.1
keyboard~=0.13.5
pynput~=1.7.6